job_matcher.py
from sklearn.neighbors import KNeighborsClassifier
import numpy as np

def match_jobs(candidate_id, job_id):
    # Simulated model and scoring
    candidate_features = np.random.rand(1, 5)  # Simulated features
    job_features = np.random.rand(10, 5)      # Simulated features

    knn = KNeighborsClassifier(n_neighbors=3)
    knn.fit(job_features, np.arange(10))  # Simulated training

    job_match = knn.predict(candidate_features)
    return {"best_match_job_id": int(job_match[0])}
