requirements.txt
Flask
Flask-Cors
mysql-connector-python
scikit-learn
spacy
